Renamed application classes to avoid ambiguity
