+++ services/hotel-svc/svc/src/main/java/org/ukma/spring/crooodle/hotelsvc/HotelService.java
@@ -1,15 +1,15 @@
-package org.ukma.spring.crooodle.usersvc;
+package org.ukma.spring.crooodle.hotelsvc;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 import org.springframework.cloud.openfeign.EnableFeignClients;
 
-@EnableFeignClients
+@EnableFeignClients(basePackages = {"org.ukma.spring.crooodle.usersvc.client"})
 @EnableDiscoveryClient
 @SpringBootApplication
-public class UserSvc {
+public class HotelService {
     public static void main(String[] args) {
-        SpringApplication.run(UserSvc.class, args);
+        SpringApplication.run(HotelService.class, args);
     }
 }
