+++ services/user-svc/client/pom.xml
@@ -23,6 +23,6 @@
     </dependencies>
     <packaging>jar</packaging>
 	<properties>
-		<svc.mainClass>org.ukma.spring.crooodle.hotelsvc.HotelSvc</svc.mainClass>
+		<svc.mainClass>org.ukma.spring.crooodle.hotelsvc.HotelService</svc.mainClass>
 	</properties>
 </project>
