+++ services/user-svc/svc/src/main/java/org/ukma/spring/crooodle/usersvc/UserService.java
@@ -1,15 +1,15 @@
-package org.ukma.spring.crooodle.hotelsvc;
+package org.ukma.spring.crooodle.usersvc;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 import org.springframework.cloud.openfeign.EnableFeignClients;
 
-@EnableFeignClients(basePackages = {"org.ukma.spring.crooodle.usersvc.client"})
+@EnableFeignClients
 @EnableDiscoveryClient
 @SpringBootApplication
-public class HotelSvc {
+public class UserService {
     public static void main(String[] args) {
-        SpringApplication.run(HotelSvc.class, args);
+        SpringApplication.run(UserService.class, args);
     }
 }
